+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -  On serialization errors the complete path to the node that triggers the
    error is printed. Only active for debug builds.
 -  Trigger server shutdown if any of the task queue workers aborts.
+-  PRACTICE functions as built-in functions with semantic token scope
+   `function.defaultLibrary`.
 
 ### Changed